% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name_variations.R
\name{get_name_variations}
\alias{get_name_variations}
\title{Get Country Name Variations}
\usage{
get_name_variations(country, country_type = c("iso3", "iso2",
  "iso_name"), collapse = FALSE)
}
\arguments{
\item{country}{A character string with standardized country names}

\item{country_type}{The stadardization type (c("iso3", "iso2", "iso_name"))}

\item{collapse}{If TRUE variations are separated by '|' else a list is returned}
}
\value{
A list of a character vector
}
\description{
Get Country Name Variations
}
\examples{
country <- c("DEU", "GBR", "HKG", "HUN", "IDN")
get_name_variations(country, country_type = "iso3", collapse = TRUE)

}
