% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_countries.R
\name{standardize_countries}
\alias{standardize_countries}
\title{Convert Country Names to ISO Format}
\usage{
standardize_countries(country, to = c("iso3", "iso2", "iso_name",
  "continent_code"), missing = c("keep", "remove"))
}
\arguments{
\item{country}{A character string with country names}

\item{to}{one of c("iso3", "iso2", "iso_name")}

\item{missing}{}
}
\value{
A character String
}
\description{
Converts a character string of country names into ISO 3166-1 Format.\cr
Available formats are:\cr
- Alpha 2: Two Letter Country Code (iso2)\cr
- Alpha 3: Three Leter Country Code (iso3)\cr
- Common Name
}
\examples{

countries <- c("DE", "Germany", "UK", "GB", "Singapore")
standardize_countries(countries, miss = "remove")

}
